Revert "Commit to Revert <3"
This reverts commit 373c0c9b60d246580e02e3e7fce66fa8d0a01a55.

+++ wServer/networking/handlers/InvSwapHandler.cs
@@ -222,11 +222,11 @@ namespace wServer.networking.handlers
                 if (!ret)
                 {
                     log.FatalFormat("Cheat engine detected for player {0},\nInvalid InvSwap. {1} instead of {2}",
-					client.Player.Name, client.Manager.GameData.Items[(ushort)packet.SlotObject1.ObjectType].ObjectId, item1.ObjectId);
+                            client.Player.Name, client.Manager.GameData.Items[packet.SlotObject1.ObjectType].ObjectId, item1.ObjectId);
                     foreach (Player player in client.Player.Owner.Players.Values)
                         if (player.Client.Account.Rank >= 2)
                             player.SendInfo(String.Format("Cheat engine detected for player {0},\nInvalid InvSwap. {1} instead of {2}",
-					client.Player.Name, client.Manager.GameData.Items[(ushort)packet.SlotObject1.ObjectType].ObjectId, item1.ObjectId));
+                                client.Player.Name, client.Manager.GameData.Items[packet.SlotObject1.ObjectType].ObjectId, item1.ObjectId));
                 }
             }
             if (con1 is Player && con2 is Player)
@@ -236,11 +236,11 @@ namespace wServer.networking.handlers
                 if (!ret)
                 {
                     log.FatalFormat("Cheat engine detected for player {0},\nInvalid InvSwap. {1} instead of {2}",
-                            client.Player.Name, item1.ObjectId, client.Manager.GameData.Items[(ushort)packet.SlotObject2.ObjectType].ObjectId);
+                            client.Player.Name, item1.ObjectId, client.Manager.GameData.Items[packet.SlotObject2.ObjectType].ObjectId);
                     foreach (Player player in client.Player.Owner.Players.Values)
                         if (player.Client.Account.Rank >= 2)
                             player.SendInfo(String.Format("Cheat engine detected for player {0},\nInvalid InvSwap. {1} instead of {2}",
-                                client.Player.Name, item1.ObjectId, client.Manager.GameData.Items[(ushort)packet.SlotObject2.ObjectType].ObjectId));
+                                client.Player.Name, item1.ObjectId, client.Manager.GameData.Items[packet.SlotObject2.ObjectType].ObjectId));
                 }
             }
